common/request_util.py
import allure
import requests


class RequestHandler:
    def __init__(self):
        """session管理器"""
        self.session = requests.session()

    @allure.step('请求接口')
    def visit(self, method, url, params=None, data=None, json=None, headers=None, **kwargs):
        return self.session.request(method, url, params=params, data=data, json=json, headers=headers, **kwargs)

    def close_session(self):
        """关闭session"""
        self.session.close()
