requirements.txt
pytest~=7.1.2
pytest-rerunfailures    #用例失败后重跑
allure-pytest  #用于生成美观的测试报告
pyyaml~=6.0
jsonschema~=4.5.1
requests~=2.27.1

